+++ src/components/favourite-item/favourite-item.module.css
@@ -1,0 +1,9 @@
+.cardImage {
+  width: 150px;
+  height: 110px;
+}
+
+.svgPlaceCardBookmarkIcon {
+  width: 18px;
+  height: 19px;
+}
